third_party/blink/web_tests/inspector-protocol/css/css-set-scope-text-reattach-expected.txt

Tests CSS.setScopeText with scope nested in a rule.

Running test: testNestedEdit
==== Style sheet text ====

  .outer {
    @scope (.b) {
      .inner { color: red; }
    }
  }
